chore: rename file to resend-dialog
